Support saving old trips with former leaders
Previously, if a trip had some leaders that are now just normal
participants, we would silently remove the participant from the list of
leaders.

That was obviously really bad, but since fixing that bug, a new (known)
one arises: the back end would not allow saving a trip where one of the
leaders was not currently an active leader.

Fix that!