Improve lottery placement when last on a trip
The lottery algorithm will currently let a driver displace a participant
who's already on a trip when that driver's attendance is necessary to
make the trip a reality: In a ten-person trip, dropping the last
participant off means the next 4 participants can now attend. Without
dropping the last participant, we'd instead have 5 participants that
can't come at all!

While this policy of dropping the last participant on a trip is
generally sensible, it can result in some frustrating edge cases.
For example, we often have a single driver who displaces one of two
paired participants (moving one to the waiting list). These paired
participants would generally prefer to be placed on another trip, but
together!

This change applies a cautious strategy. If a trip is eligible for
driver displacement, and some future unhandled drivers have expressed
interest, avoid placing a participant on that trip and instead seek
which other trips they may have selected.