Allow configurable opt-out from HIBP API
When testing locally, I tend to set user passwords to a terrible
password totally unfit for use in production. This is helpful because
all testing users have the same (bad) password. However, it's a small
nuisance to have to go through the "insecure password!" flow (since I'm
not actually going to change this dummy bad password).

What's more, it's not really the best use of HIBP resources to just
continually hit their API with a known bad password.