Fix `signup_tags.membership_active`
This method was invoking a method that didn't exist on the object!
Because Django's templates treat these failures as false, we were
rendering the "membership that's active until <date>" for all members
who just entirely lacked a membership.

This isn't the worst bug, but it's also not ideal. Fix it, and add some
basic coverage.