Enumerate & consolidate trip ineligibility reasons
This moves to an enum-based system of specifying reasons why:

- a participant is unable to attend a trip
- a participant's profile must be updated

Previously, we used dedicated methods and kept much of the
message-constructing directly in templates, where it can be tricky to
test. This is a more robust solution that should make the addition of
new trip attendance criteria a bit simpler.