Handle edge case of long-out trip
Previously, we prevented participants from ever signing up for a trip
that's more than a year out. That's admittedly not something we ever
expect to happen, but we should solve the bug all the same.

While we're at it, add a consistent interface for determining if the
participant must sign a waiver for the trip.