Hide program & trip type when not useful
When the program is "None" or the trip type is "Other," that's not
really useful information to convey. In `trip_summary` we already hide
the display for these cases.

This also fixes a small rendering bug in Gmail (& potentially other
clients): the newline between our `<li>` tag and the first `<strong>` is
counted as "whitespace" and rendered as a `&nbsp;`. Sigh. HTML in email
is the worst.