Enable logging out on pages without Angular
We used an Angular controller to make a POST to the logout route
(one cannot log out with a GET because of the various ways to trick a
browser into loading the logout route).

However, by disabling Angular on some pages, we accidentally made it
impossible to log out from those pages!

As a temporary workaround, simply give a link to the "are you sure?"
page whenever Angular isn't loaded.