Fix discrepancy on trips needing chair review
Previously, the "ApproveTripsView" was not actually restricting its
queryset to only upcoming trips. By subclassing the "UpcomingTripsView"
class, it seemed that was the case, but it was not true.

If a past trip had taken place without chair approval, it would always
populate as the first trip in need of activity chair review. This makes
for a natural conflict with two other parts of the UI:

- The count of trips needing review (shown as a badge in the menu)
- A message reading 'No current or upcoming trips!' when 0 trips

This commit fixes that discrepancy by making sure that we only count
trips taking place on the current day or in the future.

As a bonus, this also makes the view marginally more efficient by
reducing needless queries & subclassing.