Allow WS leaders to sign up with recent attendance
If a Winter School leader has recently attended lectures, we should
allow them to sign up for a trip as a participant (even if they've not
attended the current year's lectures). This carve-out should help a lot
of the leaders who miss lectures because they're helping out with office
hours or other pre-WS tasks.