Resolve ambiguity on simple trip list
The simple trip list table is used in a number of places where we want
to succinctly describe a collection of trips. The most common place is
on the participant's profile where we list their upcoming trips.

It's possible for the participant to have trips coming up on a weekend
day, but other trips on a long-distant weekend. Previously, we would
render the "date" for each of these as just "Sat" or "Sun" with no
indication about date.

Fix the ambiguity by:
- Only using day of the week when it's the closest Wed, Thur, etc.
- Using "Today" for trips leaving today
- Use a brief date for distant trips
- Apply a tooltip that gives an unambiguous datetime