Update worry time to reflect new WSC guidelines
Though not all trips are winter trips, the majority of MITOC trips will
take place in the White mountains. The same rules that have the WSC
encouraging winter leaders to set an earlier worry time should apply
year-round (though sunset will definitely be later in the non-winter
months).

If we find that winter policy differs significantly from 3-season
policy, we can change the help text to be dependent on trip type.