Hide the "let participants drop" option from WS
WS trips should always have the admins in charge of modifying the trip's
attendants. If a participant drops off a trip while it's in the "quiet"
phase before trips become FCFS, that can result in there being an open
slot that's not automatically filled by somebody on the waitlist.

Additionally, we don't want participants to be able to easily "flake"
without the leaders knowing that they dropped off.