Correct communication around itinerary opening
We used to not open itineraries for submission until Friday, but we've
recently bumped up to Thursday night. Rather than saying this in
hard-coded text, we should instead state _exactly_ when the itinerary
opens.

The next commit will introduces the concept that, once submitted, past
trips should not have their itineraries be editable.