Introduce concept of being seen vs handled
When we're going through participants in the lottery, it's helpful to
keep track of those whose number has come up (but we delayed actually
placing them). However, we don't want to conflate that with the concept
of participants being "handled" (in that they were placed and should no
longer be considered).

We don't currently support 2 drivers bumping off the last 2 people on a
trip, but if we ever do, the `_placement_would_jeopardize_driver_bump`
method was improperly checking if paired participants were _seen_, not
handled.