Add module for detecting cycles
When dealing with requested lottery separations, it's possible that some
of the separations will result in a cycle. These cycles represent an
algorithmic weakness - if each participant is waiting for another in the
cycle to be placed first, then we'll have an endless race condition.

Being able to detect these cycles and act upon them is critical. Add
a module for detecting cycles - once handling separations in the
lottery, we'll have to consciously (and cautiously) break cycles to
ensure that the lottery can come to an end.

Note: Should we ever choose to implement detection of strongly connected
components in the graph of requested signups (a useful foundation for
identifying Pareto efficient changes we can make after running the
lottery), this module would be an ideal place to put that logic!