Create concept of "desired" trips
Previously, we started with all future signups for WS trips given by any
participant, then discarded anything that didn't mean certain criteria
(for example, if they are paired, their partner must also have selected
those same trips). This made it such that ranked trip information did
not reflect what the participant actually chose.

We're going to do more with the concept of desired trips when we start
incorporating lottery separations into the algorithm. Start a framework
for it now, which we can also use to implement a participant feature
request around pairing up with a leader.