Don't show leaders a participant's feedback
Most leaders have no reason to view a participant's feedback when
viewing a profile. In some cases, though, they do! A common example is
evaluating if a given participant would be a good fit for a trip when
the participant has emailed the leaders.

To preserve participant privacy, we should make sure that leaders only
view feedback when it's necessary (and add some basic logging each time
that a leader deems this necessary).

Start with a simple, JavaScript-free version of this functionality.
We'll soon also provide a Vue-based alternative version (where
"unscrambled" comments are fetched as part of an AJAX request and the
DOM is updated without a full page reload). This functionality will be
preserved to serve leaders who use the site without JavaScript.