Remove pipeline HTML-compressing middleware
Django 2.0 dropped support of Python 2, so it can use the correct
`bytes`/`str` types freely.

`django-pipeline` does not appear to handle a bytestring
`response.content` properly -- it ends up being encoded as bytes twice!
That is, the resulting bytestring looks like:

    b'b\'...\''

This manifests itself most dramatically as newlines being escaped:
(`\\n' instead of '\n')

This change should be safe to send out as-is, because:

1. We're going to move away from `django-pipeline` soon anyway
2. With gzip compression turned on, HTML compression is less
   useful at delivering meaningful compression gains
3. We already have errors with visual spacing differing in production &
   local testing because of the HTML compression.
4. JavaScript is the largest part of page delivery. Some extra bytes in
   pure HTML is not a huge problem.