Remove GZip middleware
Running on-the-fly GZip compression when developing locally with
`runserver` will occasionally trigger an `ECONNRESET` and bubble up as a
`ConnectionResetError` on Django 2 and later. I don't fully understand
the root of this issue, but it seems to only affect the local test
server, and only when browsing with the devtools closed. (Hypothesis -
something to do with keep-alive behavior triggers this difference?)

We shouldn't be doing on-the-fly compression with Django anyway. NGINX
has this capability with `gzip on`.