Upgrade from Django 2.1 to 2.2 LTS
We were previously on 1.11, an LTS release that's slated to lose support
in April of 2020. I had prepared for a migration to 2.2, but had put off
the actual upgrade until after Winter School.

Now that dependencies all support Django 2.2 and Winter School is behind
us, we can safely upgrade to this newest LST version, which is to be
supported until April of 2022.

At present, we can't move to Django 3 without dropping a few
dependencies (which we actively *want* to be removed from the stack).
Chiefly:

- `django-angular`
- `django-pipeline`

There remain a number of "TODOs" that were written when running previous
versions of Django. Those should be addressed!