Fix test that fails due to unmocked time
WS applications' default year automatically bumps up to the next
calendar year once winter is over.

That is, if you're applying in the summer or fall of 2020, it's probably
for Winter School 2021.

It's beside the point that there almost certainly won't *be* a WS 2021,
but our unit tests should at least account for the possibility.