Use the gear rental API over direct DB access
Directly reading from another project's database is a huge anti-pattern
that we've just quietly dealt with for years.

This represents a migration away from that pattern towards one of
services communicating with one another over APIs.

This will impose a small delay in fetching results, but the decoupling
is worth it.