Use the gear rental API over direct DB access
Directly reading from another project's database is a huge anti-pattern
that we've just quietly dealt with for years.

This represents a migration away from that pattern towards one of
services communicating with one another over APIs.

This will impose a small delay in fetching results, but the decoupling
is worth it.

Note that all MITOC leaders will see a "Rentals" tab when viewing trips.
Previously, this was just rendered in-line since it was a direct
database call. Now that there's the overhead of an API, this is somewhat
problematic since:

1. It will slow the total page load
2. It will crash viewing of a trip if the API call fails