Document defense for JWT "none" algorithm attack
This commit accomplishes two things:
1. Documents that a known attack doesn't work
2. Fixes a 500 if malicious users try the known attack

Longer explanation:
===================

JWTs can be signed with a variety of algorithms. The spec requires
implementing a 'none' algorithm which does... no signing. Was it a good
idea to put in the spec? Probably not, but here we are.

The special `none` algorithm allows attackers to request data using a
JWT which is not actually signed using a secret. If an API decodes
tokens and allows the `none` algorithm, then that means it can grant
access to attackers. This is bad.

Because we're on a more modern version of PyJWT, we've not been
vulnerable to this. Add a unit test that documents this.

For more, see:
https://auth0.com/blog/critical-vulnerabilities-in-json-web-token-libraries/