Allow only HMAC 256 or 512
The default algorithm for PyJWT is HMAC 256: If you call `jwt.encode()`
without specifying an `algorithm` keyword argument, then that's what
you'll get.

A full list of supported algorithms is here:
https://pyjwt.readthedocs.io/en/stable/algorithms.html

At time of writing, that page lists:

    HS256 - HMAC using SHA-256 hash algorithm (default)
    HS384 - HMAC using SHA-384 hash algorithm
    HS512 - HMAC using SHA-512 hash algorithm
    ES256 - ECDSA signature algorithm using SHA-256 hash algorithm
    ES384 - ECDSA signature algorithm using SHA-384 hash algorithm
    ES512 - ECDSA signature algorithm using SHA-512 hash algorithm
    RS256 - RSASSA-PKCS1-v1_5 signature algorithm using SHA-256 hash algorithm
    RS384 - RSASSA-PKCS1-v1_5 signature algorithm using SHA-384 hash algorithm
    RS512 - RSASSA-PKCS1-v1_5 signature algorithm using SHA-512 hash algorithm
    PS256 - RSASSA-PSS signature using SHA-256 and MGF1 padding with SHA-256
    PS384 - RSASSA-PSS signature using SHA-384 and MGF1 padding with SHA-384
    PS512 - RSASSA-PSS signature using SHA-512 and MGF1 padding with SHA-512
    EdDSA - Ed25519255 signature using SHA-512. Provides 128-bit security

It's insecure "to mix symmetric and asymmetric algorithms because they require
different types of keys."

https://pyjwt.readthedocs.io/en/2.0.0/api.html#jwt.decode

Because we only really need HMAC, restrict allowed choices to 256 and
512.

Note that this changes the behavior of an attacker trying to sneak by an
`algorithm='none'`. Instead of raising an exception on algorithm ==
'none' with a none-null key, we now get an exception that the algorithm
is not supported.