Downgrade Jedi to deal with ipython compatibility
This only affects local development when using `poetry install
--extras=debugging`.

Specifically, we experience crashes in both `ipdb` prompts or the
`ipython` shell (which is used by `manage.py shell` if ipython is installed).

Later versions of ipython fix this, but while we're waiting for a new
release, we can just pin to an older version of jedi.