Redirect to the current page after login
When users are given a prompt to log in, we should redirect them to the
page they were on. We already do this on most places where a log in link
is shown, *except* for the one that prompts users to log in so that they
can sign up or the main log in button.

This is obviously better UX, and the kind of thing I've missed by virtue
of just always being logged in locally & on prod.

Django's built-in URL escaping prevents any injection attacks, and using
`request.path` drops any query args (which we use exceedingly minimally
anyway).

Kudos to Adam Rosenfield for the suggestion.