Remove "magic trailing comma" for Black upgrade
In Black 19.10b0, having a comma at the end of the last value in a
collection doesn't do anything special (it can be left there, or
removed, with no real change to formatter behavior).

In newer versions of Black though, this is the "magic trailing comma" -
it'll instruct Black to automatically break out values by line. Start
tweaking source now so it won't be affected by the upgrade.