Upgrade HTTP utility libraries to close vulns
The versions of `urllib3` and `httplib2` that we had installed were
vulnerable to some exploits (which I don't believe to have affected us,
but are still worth patching).

`httplib2` could experience a DoS from hitting a malicious server, and
`urllib3` omits SSL cert validation on some HTTPS to HTTPS proxies.