Upgrade `markdown2` to prevent user DoS
This closes a vulnerability which could actually be exploited
(admittedly, only by a privileged group of users).

CVE-2021-26813: if an attacker provides a malicious string that's
ultimately passed to the `markdown2` processor, they could DoS the
server.

The only place where this system parses Markdown is in trip
descriptions, and only MITOC leaders may create trips. So, the only way
one could exploit this vulnerability is to first become a MITOC leader,
then upload a bad trip description. Still, we should patch this.