Update dev dependencies, prod subdependencies
This updates all dependencies which are either only installed in the dev
environment, or are subdependencies of production dependencies.

A notable exception is `mypy`, which we're keeping pinned (updates there
are breaking, and will come in a separate commit).

The goal of this commit is to be a low-risk update to the production
environment (since I haven't been actively testing things locally, and
the amount of daily use of the trips site has been understandably
decreased in recent times).