Unify the `ws` and `auth_db` databases
This commit undoes a mistake in architecture that I've never been too
pleased with, but never had a compelling reason to fix.

Steps that can be taken after this is rolled out:

1. Add a foreign key constraint to `participant.user_id`
2. Clean up unnecessary middleware which bridges between user/par
3. Make sure that unnecessary queries aren't being run now (n+1)

Background
==========
Many years ago, I'd intentionally moved all of the core auth models out
of the main `ws` database, and into a separate db (which I termed
`auth_db`). The idea was to create a simple way for two entirely
separate Django applications (the gear management system & trip
management system) to share the same core user models, reading & writing
directly from the same db.

Even at the time, I knew this was a less-than-ideal design: it's faster
to write & faster to use than an API layer, but introduces all sorts of
other problems. Thankfully, we never actually started sharing these
models, so there's nothing stopping us from moving back to a single
database model (and the two Django applications now communicate via an
API).

Problems this solves
====================
- `participant.user_id` can soon have a foreign key constraint added
- We can drop all references to getting a cursor for `auth_db`
- There's no longer a need for a router at all
- Without a router, there's no need to drop "hints" into models about
  which database they belong to.
- Unit tests will be faster since we only need to create one database
- Unit tests no longer need to manually specify their databases
  (notable exception: tests which hit `geardb`, also being deprecated)

How to rollout
==============
**This commit is not sufficient to complete the migration.**

If anybody's developing locally, this represents a bit of a breaking
changes: Django will start trying to read `auth` models & such from the
`default` database, and won't find them. The easiest way to fix this is
to drop and re-create a database.

In production, I'm going to just impose a small amount of downtime and
copy over all the data from `auth_db` to `ws`. It's a shockingly small
amount of data, and because the old `auth_db` can stay as it was,
there's no need for prolonged downtime.