Remove all (auto-generated) Permission objects
At present, I don't really bother with Django's Permission model, and
instead rely on a more literal means of permissioning (for example, if
you have an active "leader" rating, you are therefore allowed to do
things leaders of that particular rating may do). Permissions are
granted by means of other database constructs which have inherent
meaning and are shown to end-users in a way plain permissions might not.

In the future, I could refactor this codebase to use Django permissions
in a more standard way (e.g. once you're shown to be in the "leader"
role, you therefore get the sorts of permissions that any leader would
get, but that's not needed yet).

For now, I don't need the 150 or so automatically-generated Permission
objects. Assigning these rows to users won't do anything.

This will also help moving away from the `auth_db` database, since there
are some foreign key constraints in there that would be nice to resolve.