Bump the maximum year allowed for car model
Only 6 months late on this migration, but I imagine the 2023 year model
cars aren't terribly far away. =)

It's a wee bit annoying that Django's migration model makes this SQL
no-op into a new "migration" over and over. Maybe we just do this
validation in the frontend and be done with it?