Add foreign key constraint to Participant.user_id
The `user_id` column was always representing a link to a foreign table,
it just couldn't have a constraint because that table lived in a
separate database entirely (`auth_db`).

Now that the `user_id` column should be pointing to rows in the same
database, we can add a foreign key constraint. This is the first step
towards greatly simplifying the bridge between `Participant` and `User`
models.

Note that there may be a small performance regression since we stop
pre-fetching the user's groups. But we'll optimize that again later.

In the future, we can drop things like the `from_user` method.