Don't load participant rentals by default on trips
This is a temporary measure (hopefully it doesn't last to 2022). Plan:

1. Stop hitting the gear MySQL database directly from this repo.
2. Make Django URL `/trips/{pk}/rentals` which will show rented items.
3. Drop AngularJS directives, move to Vue (& upgrade Bootstrap)
4. Make an API route for leaders to get rented items on a trip.
5. Write a new Vue component to make this API call asynchronously,
   render rented items per participant, per trip.

Background
==========

We have a handy feature which enumerates the gear that all participants
rented on a given trip. Leaders make extensive use of this feature
during Winter School, where they are interested in seeing which other
trips' participants are holding gear that could be swapped.

We recently switched to loading all rented gear items via an API call,
which is great. However, the old pattern of hitting the MySQL database
on each trip page load is not a good pattern anymore. If the gear
database were to go down for any reason, it blocks page loading
entirely.

Why use this hack?
==================
The proper fix is to make a frontend component which renders rental
items per user, making the call asynchronously (we already have an
AngularJS directive which does this for single users).

However, my first priority right now is getting rid of all direct MySQL
access. I'm also trying to get off AngularJS and onto Vue, off Bootstrap
4 and onto something newer (Bootstrap 5?). Accordingly, building a new
AngularJS directive is not a good time investment right now.

Because per-trip rentals isn't too likely to be used at all (and if it
is, it'll be by very few peopel) we can leave it hacky right now.
Including a GET param is an undocumented API so leaders who want to bulk
open a bunch of trips to see rentals have a path for that.