Stop unit testing for direct MySQL reads
This commit technically represents a regression in test coverage, but
that's somewhat okay: we're in the process of dismantling the direct
MySQL access, so the code that's being tested is not going to be
maintained (only deleted). Unit testing this integration was already
shaky at best: we don't directly control the schema at all, there's no
guarantees that unit tests will work against the real database schema.

The reason to drop this coverage is so that the test runner does not
need to spin up a MySQL server anymore. This will make the migration
from TravisCI (no longer running our builds at all) to GitHub actions a
lot easier.