Update mypy, move its config to `pyproject.toml`
mypy now supports configuration in the standardized `pyproject.toml` -
switch to it now! This will make a Dockerfile easier, as there are fewer
config files to copy over.

Note that we needed to temporarily install django-stubs from a commit
rather than PyPI (this allows us to use `pyproject.toml` configuration).