Upgrade Vue frontend dependencies
I haven't worked on the new (minimally-used) Vue frontend for a long
time, and some Dependabot alerts have been piling up since. Since this
code isn't running on a node server, and I'm not particularly concerned
about DoS vulnerabilities (it's not hard to take down a `t2.micro`),
most of the "alerts" are not high priority (or even relevant).

This article is a great write-up of why many of the node
"vulnerabilities" aren't really relevant or helpful for a simple
frontend like this one:

https://overreacted.io/npm-audit-broken-by-design/