Update django-phonenumber-field
We previously declined to update to version 2.4.0 due to some test
failures. I didn't document that choice well, but I *believe* I chose to
hold off on upgrading due to the faulty test fixed in commit
2a0fe858f8a93d4f4850d223cf2ccbe07785f010

Version 2.4.0 introduced aggressive validation, but version 4.0.0
removed it. We should be safe to move forward past this.

From the changelog for `django-phonenumber-field`:

> Starting with 2.4.0 we added very aggressive validation, which raised
> ValueError for invalid numbers. This caused problems in unexpected
> places (like when filtering a queryset). Starting with 4.0.0 we
> acknowledge that we can not completely prevent invalid numbers from
> entering the system.

https://github.com/stefanfoulis/django-phonenumber-field/blob/main/CHANGELOG.rst

Per the above, the only breaking change for 4.0 is to not "raise
ValueError on invalid phone numbers."

The breaking changes for version 5.0 do not affect us:

> - Drop support for end-of-life Django 1.11 and 2.1.
> - As the Babel dependency is now optional, you must now install it to
>   use PhoneNumberPrefixWidget. If you do not install it, an
>   ImproperlyConfigured exception will be raised when instantiated.

By updating, we now also get compatibility with Django 3.1

Lastly, explicitly add the `phonenumberslite` extra. We already install
that dependency as a top-level dependency, but the model fields from the
django-phonenumber-field package won't work without it, so specify that
plainly.