Serve Google Form for climbing leader applications
The new climbing leader application is quite long -- up to 40 questions,
depending on what ratings a prospective leader is seeking.

The questions on the form are conditionally required as well; if you're
not seeking a bouldering rating, you don't need to answer any questions
about spotting, pads, etc.

Accordingly, a dynamic form is much better suited to handle this
application than our simplistic setup for managing applications in-app.
Use a Google Form for new applications, but preserve a record of old
applications for now.