Fix display of leader applications w/ no chair
This fixes a bug for displaying applications when there is not an active
activity chair. In such a scenario, only an admin user (read: me) will
actually be able to see applications. In any case, we shouldn't show a
message which says "you're given a recommendation for all applications"
because that's not actually true.

Instead, we should regard all applications as needing a rating.
Presumably, chairs will step up and/or be elected in the future, and can
then take care of either making recommendations to one another or just
assigning ratings.