Add the option to complete an embedded form
This adds the option for those who are applying to be a climbing leader
to do so directly from the site (in an iframe). This can be handy, but
is also not required.

Notably, this is the first `<iframe>` used on the site. Our strict CSP
will prevent iframes by default, so we want to be sure that this one is
allowed. The CSP header must be changed from 'none' to include:

```
frame-src https://docs.google.com/forms/d/e/1FAIpQLSeWeIjtQ-p4mH_zGS-YvedvkbmVzBQOarIvzfzBzEgHMKuZzw/;
```