Upgrade pylint to latest release, 2.10
In Python 3.10, calling `open()` without the `encoding` argument will
raise an `EncodingWarning`. This is to help an eventual switch to making
`UTF-8` the default encoding, but also to help cross-platform
compatibility:

> Developers using macOS or Linux may forget that the default encoding
> is not always UTF-8.
> ...
> Many Windows users cannot install such packages if there is at least
> one non-ASCII character...

https://www.python.org/dev/peps/pep-0597

The major change present in pylint 2.10 is to lint any calls to `open()`
that don't explicitly state their desired encoding.

This software is only meant to run on POSIX-compatible systems, but all
the same - it's good to be explicit about encodings.