Update to Python 3.8
It's high time that we upgrade past Python 3.6. Python 3.6 will be EOL
on December 23, 2021. Python 3.8, in contrast, is supported until the
end of 2024. We also get a number of new features that are presently
backported via a number of dependencies.

https://devguide.python.org/#status-of-python-branches

This also lets us run the project with the default Python version on Ubuntu 20.04
(Focal Fossa). In the future, we ideally will install & manage our
Python versions in a container rather than using system Python.

Notably, a *lot* of Python packages are now removed. This slimmer
dependency tree comes from the fact that a lot of backports (like
`dataclasses` and `typed-ast`) are no longer needed.