Use stdlib datetime instead of custom methods
On Python 3.6, we were using the library `dateutil` to provide
relatively simple functionality of parsing an ISO 8601 date or datetime
into a Python `date` or `datetime` object.

Now that we're on Python 3.8, the standard library can answer these
questions, and there's no more need for our helper methods. We left the
helper methods in place for a brief while to demonstrate that
functionality had not changed when moving to the stdlib implementations.

With 2b33eb27ad924b2714676e9a9981b63cfa0398bb passing the build, we can
confidently do a bulk replacement of everywhere those methods were used.