Remove the obsolete test config
This configuration was used for a little while to disable signals in
unit tests. In hindsight, using signals at all was probably a mistake
(they're a fraught pattern). In any case, we haven't used the TestConfig
since 2017, when it was removed with 0e97aa09b95ced503b05f994f2bc675adb1a33f8