Install Django crispy forms
We have a number of forms that are tightly coupled to AngularJS, relying
upon the framework for frontend validation, and defining forms to
specifically output HTML attributes to bridge the gap. This is a pattern
I'd like to move away from as much as possible. Having your backend &
frontend coupled is of course a problematic pattern in and of itself,
but AngularJS is a deprecated framework.

However, one perk of the existing forms is that they automatically style
forms for use with Bootstrap 3. We should preserve this automatic
presentation. Installing Django Crispy Forms (a purpose-built and
well-maintained framework) will help us keep using Bootstrap 3 form
layouts, decouple direct AngularJS integration, *and* set us up to get
onto Bootstrap 4 or even 5.