Use Bootstrap rendering on basic auth forms
This corrects a long-standing presentational complaint: forms used for
auth were poorly styled & inconsistent with the rest of the site. Now
that we have Crispy Forms installed, we can render forms with standard
Bootstrap 3 styling, which is consistent with the rest of the site.

The default styling here is courtesy of the setting
`CRISPY_TEMPLATE_PACK`.