Remove Django-Angular from profile form
The participant profile page is the primary place where we make
productive use of Django-Angular: we use Angular to collect phone
numbers via a handy AngularJS widget, and we use Angular form validation
to request that users provide a full name.

This change ensures that we can still have dynamic behavior in the
frontend (presently, with AngularJS), but that we no longer bother with
Django-Angular. We're not really losing much, since the only real
validations that are non-trivial are:

- Valid phone number validation (immediately rendered in red)
- Full name validation (now done with HTML5)
- MIT status with non-MIT email (plainly rendered if mismatched)

We can now allow submission to the backend & let the backend provide
error messages. This makes more a much better experience when browsing
without JavaScript. We remove `novalidate`, letting native input
elements do client-side validation with no JS!