Remove use of the Django-Angular URL middleware
This commit removes usage of `djangoUrl`, an AngularJS helper which
makes it so that an AngularJS application can reference *named* URL
routes rather than creating the URL directly.

We remove the method primarily because the it doesn't solve a serious
problem, but also because removing the URL translation gets us one step
closer to deleting `django-angular` from the codebase.

In hindsight, using helper this was never a particularly good idea. Once
decided upon, URLs shouldn't change. Per W3C:

> Cool URIs don't change

https://www.w3.org/Provider/Style/URI,

However, assume that you're *not* cool and must change a URI's name
within Django. The `djangoUrl` method does *not* help you keep the
frontend and backend in sync: if a URL definition is edited in
`urls.py`, stale JS bundles cannot use its old name; they'l 404.

We thankfully don't need to manage the pitfalls of a SPA, where users
can continue to hit many API endpoints before getting a fresh bundle.
However, users still can interact with a stale page which calls
`djangoUrl()` using a since-renamed method.

So, what's the point? Why not just build a URL directly? Honestly, I
can't say, which is part of why we're removing this.