Switch leader form to plain ModelForm
This is migrating yet another form away from Django-Angular and towards
a more barebones alternative. We use Crispy Forms to keep the Bootstrap
3 styling of form fields, but just pass down an `ng-model` specifically
for the field we wish to inspect.

Add some basic test coverage while we're at it.

These tests are only the most basic functionality, we don't cover:

- rendering things like feedback, past trips, etc.
- handling existing recommendations
- other applications being present
- the prev/next functionality
- leaders who already have ratings

These tests should be added in the future