Restore functionality on "Create/Update Ratings"
This form was still functional from a pure HTML/API standpoint, which is
why our tests didn't catch the change (we really need integration tests
at this point).

In commit 7f2a73a075b2105f065f4bcf84b0db46a9c52f4e, I had changed a few
simple forms to be rendered with Crispy Forms instead, but I
accidentally committed changes to this form too without addressing the
use of AngularJS in the `leaderRating` controller.

Of course, in the future we should just use VueJS, but we also need a
new participant selector that isn't based in Angular.

For now, this restores the functionality that when you look up a
participant's name, ratings for any given activity pre-populate the
form.