Fix the form available at /chair/leaders
This change is done as a follow-on to the previous commit,
7f8ba1b620cdbc8d44e000219c93dd7570e3b5b7 - I changed the form to no
longer be based on Django-Angular, so it now lacks the `as_div` method
that renders it.

Fix this by switching to rendering with Crispy Forms, and also make the
form narrower on large screens.

However, this form is never really used. Most activity chairs just
manage leaders for their own activity, and that's all.

I'll keep this page for now, since it's harmless, but in a future
refactoring I should strongly consider removing the page, and changing
the LeaderForm to take its `activity` from URL params, not a form
widget.