Add tests for leader signup
There was a latent error in the leader signup route where a
non-specified trip would cause a 500 (we check for notes, but fail to
find a trip).

Additionally, upon inspection there's not really anything preventing a
leader from being a participant on a trip and then *also* signing up as
a leader.

We should fix both of these issues, but first start by adding some test
coverage that can form the basis of future fixes.