Enforce that leader signups are allowed
Before this commit, the only thing that was technically stopping an
(appropriately-rated) leader from signing up for any trip, as a leader
on that trip, was the UI conditionally rendering the signup form.

This commit now adds some protection at the API layer so that we don't
accidentally permit this. It's not likely that a MITOC leader would
exploit this maliciously, but it is possible that a trip's settings are
altered so that leader signups are no longer allowed & some stale
rendered pages come in with new leader signups.