Fix a possible 500 on trip signup
If submitting a signup for a trip, but the POST body does not include
the trip ID for any reason, we would previously 500 when trying to check
the notes on the signup (since `'trip'` was not present in the cleaned
data).

This fixes that erroneous assumption & adds some tests which would have
failed before the fix.