Remove the top-level COVID warning
MITOC is now running some in-person events. We're certainly not back to
"normal," and MIT guidelines of course still apply. However, the notice
that all events are canceled until further notice is no longer
applicable.

Do continue to warn on the membership payment page that the office
remains closed: some members have paid their dues, and been confused to
see that the office is still not open to the public.