Stop using Django-Angular on membership form
The membership form has to work in three different contexts:

- Participants with a known affiliation
- Anonymous users without a known affiliation
- Anonymous users who have JavaScript disabled

We don't need Django-Angular for any of these: we can just use plain
AngularJS to supplement the form (for users who do indeed have JS turned
on).

I tested the form in each of these three environments, and can confirm
that basic browser validation is still working when JS is disabled.