Rename MITOC Rock Program back to School of Rock
Initially, the program was known as SoR, but it was briefly renamed to
MITOC Rock Program. We're now back to referring to this as School of
Rock.

We can't easily update the value that we use in the database (though a
migration may simplify that), though we *can* change the enum name.