Change the lottery pairing to a simple form
This is now the second to last form to still use the legacy
Django-Angular framework.

This is a particularly simple form. We set some attributes on the select
widget itself, but there's no real need for validation: any participant
pk is accepted, as is an empty selector (no participant).

Also, we should remove the pointless `exclude_self` kwarg: you can
*never* pair with yourself, and the queryset is defined to restrict
choices down to participants who are not you.