Rewrite how we do lottery preferences
This commit represents both a reorganization of the lottery preferences
page as well as a few coding improvements. It also accepts a bit of a
regression in functionality to support future refactors.

Reorganize the page
===================
Some messages have been moved around, spacing adjusted, etc. There's a
lot going on with this page, and trying to (succintly) inform the user
the minimum they need to know is worthwhile.

Outside of Winter School, the right sidebar can be repurposed to show
information about lottery pairing, since trip ranking won't happen
there.

Stop collecting car information on the page
===========================================
As is, the process of collecting car information at the time of the
lottery preferences is a weird pattern. Some people have access to
different cars, or might change what car they drive for a given trip.
Our system naively assumes that car owners have just one car, and
they'll only drive that one car week after week. From a process
standpoint, these assumptions are not good.

Furthermore, the technical way we've implemented car submission from
lottery preferences -- submitting a JSON payload representing a car then
saving a form from that -- is very much not a good Django pattern.

Replace this all, and go with a really simple version - where we prompt
users to go to their profile if they need to edit their car.

This may have downsides, namely that users say they can drive a car, but
never give details. Hopefully we fix that by WS 2022 (if it happens).
Ideally, we'll want to add the ability to manage *multiple* cars per
participant, so that should be the goal.

Pass JSON from backend to frontend better
=========================================
We want to supply a list of ranked signups from the backend to be
consumed by the AngularJS frontend, without the need for API calls.

In the future, we could instead opt for an API call (probably ideal when
refactoring to use VueJS). However, we can at least now stop inserting a
`<script>` take with executable code, and use the safer (XSS-resistant)
`json_script`. This will help enforce a stronger CSP.

Warn participants who only sign up for a few trips in WS
========================================================
It doesn't happen often anymore, but a few participants only sign up for
one or two trips, and are disappointed when they are placed on no trips.
We should give a specific warning for anybody who selects 3 or fewer
trips.

Experimentally, there's a huge cutoff in placement likelihood between
participants who selected 3 or fewer trips and those who selected more.

Drop Django-Angular
===================
The form we used for lottery preferences was the last to use
Django-Angular! We can clean up dependence on this package now.

Additionally, we remove an old dependency from the AngularJS
`lotteryController` that was breaking form submission. This was an
incomplete cleanup from 27b9e70cb4f9b022a0daf78deed48ffbfb245c06